refactor: move `health_check` to submodule top level
